backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
google-generativeai==0.5.4
httpx==0.27.0
pymongo==4.7.2
python-dotenv==1.0.1
pydantic==2.7.1
Pillow==11.0.0
numpy
tensorflow
passlib[bcrypt]
pyjwt
